langserver: Move listPkgUnderDir to internal tools pkg
We will start using this pkg more for storing functionality that is closely
related to code from golang.org/x/tools
