+++ langserver/internal/tools/doc.go
@@ -1,0 +1,3 @@
+// tools mainly contains variations on code found under golang.org/x/tools,
+// but modified to be optimized for our usecase.
+package tools
